Adding search button on navgation bar on ou selection component
