+++ src/components/organisation-unit-tree/organisation-unit-tree.html
@@ -5,7 +5,7 @@
     item-left></ion-icon>
 
   <div (tap)="setSelectedOrganisationUnit(organisationUnit)">
-    <span ion-text [color]="(shouldIndicateAssigmentsIssues && ouIdsWithAssigments && ouIdsWithAssigments.indexOf(organisationUnit.id)  == -1 )? 'ou-assigment-issues' : (currentSelectedOrgUnitName == organisationUnit.name) ?'current' : 'font-color'">{{ organisationUnit.name }}</span>
+    <span ion-text [color]="( ouIdsWithAssigments && ouIdsWithAssigments.indexOf(organisationUnit.id)  == -1 )? 'ou-assigment-issues' : (currentSelectedOrgUnitName == organisationUnit.name) ?'current' : 'font-color'">{{ organisationUnit.name }}</span>
   </div>
 
 </ion-item>
@@ -31,8 +31,8 @@
   <div *ngIf="isOrganisationUnitsFetched && hasOrgUnitChildrenLoaded && !hasErrorOccurred">
     <div *ngFor="let child of organisationUnit.children">
       <organisation-unit-tree [organisationUnit]="child" [currentSelectedOrgUnitName]="currentSelectedOrgUnitName" [ouIdsWithAssigments]="ouIdsWithAssigments"
-        [shouldIndicateAssigmentsIssues]="shouldIndicateAssigmentsIssues" [hasOrgUnitChildrenOpened]="hasOrgUnitChildrenOpened"
-        (selectedOrganisationUnit)="setSelectedOrganisationUnit($event)" [currentUser]="currentUser"></organisation-unit-tree>
+        [hasOrgUnitChildrenOpened]="hasOrgUnitChildrenOpened" (selectedOrganisationUnit)="setSelectedOrganisationUnit($event)"
+        [currentUser]="currentUser"></organisation-unit-tree>
 
     </div>
   </div>
