+++ src/pages/organisation-unit-selection/organisation-unit-selection.scss
@@ -1,3 +1,1 @@
-page-organisation-unit-selection {
-
-}
+page-organisation-unit-selection {}
